updateWithReducer -> toggleLayer
